requirements.txt
Django==1.7.1
argparse==1.2.1
distribute==0.6.24
django-braces==1.4.0
django-cors-headers==0.13
django-extensions==1.4.6
django-oauth-toolkit==0.7.2
django-tastypie==0.12.1
oauthlib==0.7.2
psycopg2==2.5.4
python-dateutil==2.2
python-mimeparse==0.1.4
pytz==2014.9
six==1.8.0
wsgiref==0.1.2
